refactor: move the degraded retry schedule out of the hook that owned it
useProviderStatus needed the same exponential backoff, and importing it
from useHermesModelOptions made every test that replaces that hook
wholesale throw an unhandled rejection the moment a checking row appeared -
three component suites were one fixture away from it.

The schedule has no React and no imports, so it lives in
src/lib/degraded-retry.ts now, shared by the model catalogue, the chat
model pill and the providers panel. useHermesModelOptions re-exports it so
ChatPopup's import path is unchanged.
